+++ tests/resources/test_support/constant_low.py
@@ -152,8 +152,16 @@ FAILED_RESULT_DEFECT = {
     "enabled": True,
     "fault_type": FaultType.FAILED_RESULT,
     "error_message": "Device defective.",
-    "result": ResultCode.FAILED,
+    "result": ResultCode.QUEUED,
     "target_obsstates": [ObsState.IDLE],
+}
+
+FAILED_RESULT_DEFECT_EMPTY = {
+    "enabled": True,
+    "fault_type": FaultType.FAILED_RESULT,
+    "error_message": "Device defective.",
+    "result": ResultCode.QUEUED,
+    "target_obsstates": [ObsState.EMPTY],
 }
 
 SDP_BACK_TO_INITIAL_STATE = {