+++ tests/tmc/test_assign_resources_command_not_allowed_propagation_low.py
@@ -110,7 +110,7 @@ class TestAssignCommandNotAllowedPropagation:
         ).within_timeout(TIMEOUT).has_change_event_occurred(
             central_node_low.subarray_node,
             "obsState",
-            ObsState.FAULT,
+            ObsState.EMPTY,
         )
 
     @pytest.mark.SKA_low
@@ -193,7 +193,7 @@ class TestAssignCommandNotAllowedPropagation:
         ).within_timeout(TIMEOUT).has_change_event_occurred(
             central_node_low.subarray_node,
             "obsState",
-            ObsState.FAULT,
+            ObsState.EMPTY,
         )
 
     @pytest.mark.SKA_low
@@ -280,5 +280,5 @@ class TestAssignCommandNotAllowedPropagation:
         assert wait_and_validate_device_attribute_value(
             central_node_low.subarray_node,
             "obsState",
-            ObsState.FAULT,
+            ObsState.EMPTY,
         )