+++ tests/tmc/test_verification_skb-985.py
@@ -39,7 +39,7 @@ from tests.resources.test_support.constant_low import (
     RESET_DEFECT,
 )
 
-TIMEOUT = 30  # seconds
+TIMEOUT = 100  # seconds
 
 
 @pytest.mark.post_deployment
@@ -70,6 +70,7 @@ def given_subarray_ready(
     """
     # Subscribe to events needed later
     event_tracer.subscribe_event(subarray_node_low.subarray_node, "obsState")
+
     event_tracer.subscribe_event(
         subarray_node_low.subarray_node, "longRunningCommandResult"
     )
@@ -152,6 +153,42 @@ def wait_for_fault(
     event_tracer: TangoEventTracer, central_node_low: CentralNodeWrapperLow
 ):
     """Block until the Subarray reports ObsState.FAULT."""
+    assert_that(event_tracer).within_timeout(
+        TIMEOUT
+    ).has_change_event_occurred(
+        central_node_low.subarray_node, "obsState", ObsState.FAULT
+    )
+
+
+@given(
+    "the Subarray transitions to observation state ObsState.FAULT"
+    " from RESOURCING"
+)
+def wait_for_fault_resourcing(
+    event_tracer: TangoEventTracer, central_node_low: CentralNodeWrapperLow
+):
+    """Block until the Subarray reports ObsState.FAULT."""
+    assert_that(event_tracer).within_timeout(
+        TIMEOUT
+    ).has_change_event_occurred(
+        central_node_low.sdp_subarray_leaf_node,
+        "sdpSubarrayObsState",
+        ObsState.RESOURCING,
+    )
+    assert_that(event_tracer).within_timeout(
+        TIMEOUT
+    ).has_change_event_occurred(
+        central_node_low.mccs_subarray_leaf_node,
+        "obsState",
+        ObsState.RESOURCING,
+    )
+    assert_that(event_tracer).within_timeout(
+        TIMEOUT
+    ).has_change_event_occurred(
+        central_node_low.csp_subarray_leaf_node,
+        "cspSubarrayObsState",
+        ObsState.FAULT,
+    )
     assert_that(event_tracer).within_timeout(
         TIMEOUT
     ).has_change_event_occurred(
@@ -249,6 +286,15 @@ def given_subarray_empty(
         subarray_node_low.subarray_node, "longRunningCommandResult"
     )
     event_tracer.subscribe_event(
+        subarray_node_low.sdp_subarray_leaf_node, "sdpSubarrayObsState"
+    )
+    event_tracer.subscribe_event(
+        subarray_node_low.csp_subarray_leaf_node, "cspSubarrayObsState"
+    )
+    event_tracer.subscribe_event(
+        subarray_node_low.mccs_subarray_leaf_node, "obsState"
+    )
+    event_tracer.subscribe_event(
         central_node_low.central_node, "longRunningCommandResult"
     )
     event_tracer.subscribe_event(
@@ -264,6 +310,9 @@ def given_subarray_empty(
                 "longRunningCommandResult",
                 "telescopeState",
             ],
+            subarray_node_low.sdp_subarray_leaf_node: ["sdpSubarrayObsState"],
+            subarray_node_low.csp_subarray_leaf_node: ["cspSubarrayObsState"],
+            subarray_node_low.mccs_subarray_leaf_node: ["obsState"],
         }
     )